requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
Django==2.2.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.0
django-filer==1.7.0
django-hitcount==1.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-rest-auth==0.9.5
django-tinymce==2.8.0
djangorestframework==3.11.0
easy-thumbnails==2.7
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.7
whitenoise==5.0.1
wrapt==1.11.2
